refactor(job): remove useless Start and Stop functions from AutoDeleteDepletedClientsJob (#102)
* feat: Add auto-delete depleted clients job

This commit introduces a new cron job that automatically deletes clients who have exhausted their traffic or whose subscription has expired.

The feature is configurable via a new setting in the panel, allowing administrators to specify the number of days after depletion/expiration when a client should be removed.

* feat: Improve auto delete job and add translations

- Change auto delete setting from minutes to days.

* refactor(job): remove useless Start and Stop functions from AutoDeleteDepletedClientsJob

The Start and Stop functions in AutoDeleteDepletedClientsJob were creating a new cron instance, which was redundant as the application already has a global cron scheduler. This commit removes these functions and their usage, and the job is now added to the main cron scheduler.